Refactoring of OneForAll (in progress)
