+++ front-end/app/pong/normal/components/Paddle.tsx
@@ -2,7 +2,7 @@ import { useFrame } from "@react-three/fiber";
 import { MutableRefObject, forwardRef, useContext, useEffect, useRef } from "react";
 import { Mesh } from 'three';
 
-import { PongContext } from '../PongProvider';
+import { PongContext } from '../../PongProvider';
 
 interface Paddle {
 	keyMap: { [key: string]: boolean };
