+++ front-end/app/pong/page.tsx
@@ -1,14 +1,14 @@
 import { title } from "@/components/primitives";
 import PongScene from "@/app/pong/normal/PongScene";
 import OneForAllScene from "./oneForAll/OneForAllScene";
-import { PongProvider } from "./normal/PongProvider";
+import { PongProvider } from "./PongProvider";
 
 export default function PongPage() {
 	return (
 		<div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '80vh' }}>
 			<div>
 				<PongProvider>
-					<PongScene />
+					<OneForAllScene />
 				</PongProvider>
 			</div>
 		 </div>
